Add derivation fields to treefile, map origin → treefile
The goal here is to move away from "origin" files and
centralize on "treefile" YAML as our standard declarative
representation.

To do this, add the key "derivation" fields to the treefile,
but error out if they're actually used on the compose side for now.

Eventually of course...there's nothing stopping us from supporting
`from:` on the compose side too.  But one step at a time.

In order to actually enable this, add `RPMOSTREE_ORIGIN_TEST_ROUNDTRIP`
to the daemon environment.