ci: Use buildpod for composes and skip installing in cosa
A major part of the buildroot CI change was to decouple the FCOS
buildroot from cosa's. We're breaking that here by trying to install our
newly built rpm-ostree RPMs into cosa. So let's drop that.

Obviously it'd be good to get that coverage, though we should be covered
by our own compose tests. Maybe long-term we should just rebuild for
*both* cosa and the FCOS buildroot if they differ. (This will be
especially relevant if we ever want to add e.g. upstream CI testing for
the `next-devel` or `rawhide` streams.)

Similarly for the compose tests, we should use the buildPod, which is
the same userspace in which we built rpm-ostree.