core: Prevent installing mismatching kernel-devel
Add a little piece of logic to the depsolving to exclude all
kernel-devel packages which have an EVR not equal to the EVR of the
installed kernel-core package (if one exists).

This should help us to remove kernel-devel from the base image of
Silverblue without worrying that we get the wrong one later on during
layering operations by the enduser.