core: Don't load host rpmdb when doing composes
I was trying to do an RHCOS build using a custom rpm-ostree build
in a f34 container without having used `--enable-sqlite-rpmdb-default`.

This immediately failed trying to load the host rpmdb.
Now, in practice we don't care about this problem, and it doesn't
solve https://github.com/openshift/os/issues/552 because
the f34 rpm still can't write bdb and so just dies later.

But, I think this is the right thing to do - we shouldn't pollute
the sack with the host (container) rpmdb in the `compose tree`
case.