cliwrap: Add initial yum/dnf wrapper
I've come to the conclusion that we need to be friendlier
to people who are transitioning from existing `yum/dnf`
systems.  Just saying `yum: command not found` is harsh.

And moreover, there are some things that IMO *should* work.
Let's take as an example: https://tailscale.com/download/linux/fedora
This is exactly an operating system extension; it doesn't
make sense in `toolbox` or `flatpak`.  (It *could* be containerized
but...let's leave that aside)

So following that, this provides a useful bit of information for
the `dnf install tailscale` part that will point people instead to
`dnf install-osextension tailscale`.  I think the basics of `dnf config-manager` should be
implemented here too.

Obviously, there's a *lot* more that could be done (and changed) here.  This
is just an initial starting point for discussion.  It's not
enabled by default.

The goal is to have people be able to run `rpm-ostree deploy --ex-cliwrap=true`
and try it out.