Move reboot handling to client
Having the daemon start the reboot means that it can be killed
by `SIGTERM` before the client processes the completion, causing
the client to exit with a spurious error.

Instead, have the client start the reboot.  This is cleaner
in numerous ways.  An additional benefit for example is that
if there are systemd reboot inhibitors present, it will be
the client that blocks, not the daemon.