daemon: Remove support for session bus
We haven't actually tested this in a long time, it's just
cruft now.  Bigger picture we need to make it as ergonomic and
fast as possible to test in VMs.

(I think it likely would be worthwhile at some point having rpm-ostree run in a "stock
podman container" w/o systemd but that's a whole lot of work.)

Prep for some work on socket activation.