Use rpm-oxide to pre-validate RPMs
There are a ton of memory safety issues in librpm that turned
up when QubesOS paid for an audit of their code.  librpm is an old
C codebase using hand-rolled serialization, so it's not
surprising it has a ton of issues.

There are some attempts to fix it, but it's not going to happen
quickly.

In situations like this, the pattern of using Rust code as a
"pre-validator" works really well.  In theory, we could entirely
rely on rpm-oxide for GPG + payload checksum validation, but
in order to be 100% sure we aren't regressing security, add
it as a check in front before librpm sees the content.

While I'm marking this as
Closes: https://github.com/coreos/rpm-ostree/issues/2836
there's still more to do - I think we want to do the full
"rpm canonicalization" once that ends up in the library.