origin: Switch baserefspec → refspec when de-layering
The idea originally was that we'd switch to `baserefspec` when
performing changes such that `ostree admin upgrade` would no longer
work.  But that has long since been buggy since the introduction
of things like `rpm-ostree initramfs --enable` etc.

The only goal of this PR is to avoid "leaking state" across
origin switches, so that the work to map origin ⇔ treefile
can also (correctly) be stateless.