builtin/tree: make the postprocessing step explicit
This split some tree logic into two separate steps, a first one
installing content to tree and a second one massaging it through
post-processing logic.
The separation of concerns should be help code reading and may
make it easier to expose it directly as a CLI verb in the future,
if some consumers have a need to call into this.