daemon/deploy: Handle resetting inactive replacement overrides by name
Our helper name <--> nevra code wasn't handling inactive replacement
overrides correctly. Fix this by also including mappings from the origin
file.

This is also prep for the new code in #2863 which won't work correctly
without it because it won't be able to auto-replace an inactive override
with an active one.