builtin/tree: explicitly set _dbpath macro in global libdnf initialization
This ensures that the `_dbpath` macro is always in all dnf contexts,
in order to use the rpmdb under /usr.
Setting it explicitly makes sure that there is no logic implicitly
relying on host configuration. This is particularly relevant in
server-side compose cases, which are more likely to be running in
an environment without rpm-ostree macro file.