core: Neuter /usr/sbin/sss_cache during compose/layering
This fixes the immense spam of errors from `useradd` trying
to refresh the sssd cache in an `rpm-ostree compose tree`.
It doesn't work when sssd isn't running (as it won't be in a non-systemd container).

xref https://pagure.io/SSSD/sssd/pull-request/3959

I named the struct generically as I'd like in the future to
generalize this code to also do the systemctl wrapper, etc.