test-installroot: Validate we don't have /var/lib/rpm
osbuild is hitting issues where `/var/lib/rpm` is showing
up in the target ostree commit.  I am not reproducing it
right now, but let's validate it isn't happening in our CI.

Also while we're here, add a check that we're not using fuse
for `/var/tmp` (as happens in `toolbox` by default) because
that breaks us trying to set `security.selinux` xattrs.