passwd: reduce the amount of log messages for new entries
This collapses the logging of new user/group entries into a single
message, in order to reduce the amount of log lines that are
emitted at compose time.
This is especially relevant for scratch composes, where all user and
group entries are net-new ones and in aggregate may result in a wall
of text.
It also fixes a typo for group entries, which were previously miscategorized.