tests/vm.sh: New script to make spawning test VMs easier
This is a helper script I've had locally for a while that I cleaned up a
bit. It duplicates `cosa run` and some of `libvm.sh` which is originally
why I didn't push it. It sets up e.g. the SSH config and SSH keys just
right so that one can `make vmsync` afterwards.

Long-term I think we should fold that special sauce in `cosa run` with
switches and then wrap it around `make vmsync SPAWN=1` or something. But
for now, this should be helpful for other contributors.