cxxrs: bind OstreeRepoTransactionStats
This adds bidirectional binding for `OstreeRepoTransactionStats`.

As this type is a plain C struct (not wrapped by GObject), it
also requires tweaking the auto-binding macro.
In particular, the intermediate pointer is now "const" in order
to cover this kind of cases which lack interior mutability. The
qualifier is anyway discarded by the final transmute.