daemon: If we encounter a startup error, set it as our unit status text
This is a much smaller patch related to
https://github.com/coreos/rpm-ostree/pull/2932
which is about ensuring the client gets the actual errors instead
of generic DBus activation failure.

The obvious step then would be to have the client ask systemd
for the status text, but that's slightly messy right now and
I think I'd anyways prefer to go the socket route.

But I think this is a useful pattern; basically systemd's `StatusText`
facility is useful.