Stabilize `apply-live` flag for deploy, use in `rpm-ostree install -A`
Since we stabilized the CLI interface `rpm-ostree install -A`, we need
to stabilize the DBus interface as well.  This hence
Closes: https://github.com/coreos/rpm-ostree/issues/2875

Further, by moving this to one transaction we can drop the
incorrect "Run systemctl reboot" line.  While we have the patient
open, also change it to clarify in the non-live case that the
changes are queued.