core: Remove skip-sanity-check treespec support
As far as I can tell, nothing ever actually set this.  I think
I added it out of an excess of conservatism, perhaps intending
that the flag could be set somewhere.

But because the treespec is an internal only API, it can't be
set.

We really do always want these checks, so just delete the code
that tries to support disabling them.