core: Split out low level creation method
The extensions code is doing something that's not a compose or
a client operation (i.e. it's not doing treefile -> ostree commit).

Another way to look at it is it's using this as a thin wrapper
on top of raw libdnf.

Split off a low level API to create a context and separately
set the cachedir.

This is prep for removing treespec.