Allow rebase to OSTree commits in container images
This adds support for providing a container image as the argument for
`rpm-ostree rebase` by teaching rpm-ostree to use ostree-rs-ext's new
import OSTree commits from container images feature.
`deploy` while based on a container image is not yet implemented.
Currently, a tag for the target container image must be provided; i.e.
omitting the tag will not be a shortcut for `latest`.