upgrader: Generate "computed" origin
This is is part of reducing internal complexity - in particular
our "representations of state".  The code here is basically
printing a warning if a requested package is already in the base
tree, and avoid passing that package down to the core.
To accomplish this original code here was doing:

origin -> (computed state) -> treespec

Now that we dropped treespec in a recent PR, it's going:

origin -> (computed state) -> treefile

Instead of having anonymous (computed state) we can
just make that a new (filtered) origin file.  IOW
we now do:

origin -> origin -> treefile

This is easier to understand and debug.  Note
how duplication around checking "requires local
assembly" between the upgrader and origin code
just drops out!