rust/origin: Don't mutate input origin, do translate override
And here's where Rust's `&` vs `&mut` would have saved us
in Rust-native code.  Our translation code here was mutating
the input, which broke things when trying to use it in
the upgrader code.  But because dealing with `&/&mut` with
GLib would have been too hard, the GLib bindings accept `&`
but still support mutating the underlying object.

Further, we don't want to remove all the "transient" state
here anyways.  The history of ostree and origin files is tangled.
We just want to drop `libostree-transient` (ostree should
have an API for this).  Or stated positively, we do want
to translate the `override-commit`.