build-sys: Switch compile flag to --enable-bdb-rpmdb-default
The initial motivation here is a convenience for people building locally (me)
outside of RPM builds; forgetting to pass `--enable-sqlite-rpmdb-default`
on Fedora 34+ results in basically an unsupported state.

But since sqlite is the new default, let's be forward looking
and make the legacy case require opt-in.  This way we e.g. don't
need to deal with RHEL9 conditionals either.