core: Log repository solvables
I'm trying to debug a failure building RHCOS where no packages
are found.  I wanted the `n_solvables()` from the repos, which
I got via adding a debug print, but let's surface this information
by default.

Note to do this we need to move the printing to after we've loaded the sack.

This code also duplicates `journal_rpmmd_info()` but that's not
a new problem.