upgrade: Check imgref digest before pulling image
Previously, `rpm-ostree upgrade` was re-pulling a container image
each time, and throwing it away if it realizes that the commit is
the same as the currently booted deployment much later on.
Be smarter about this by checking the image reference digest, and
don't bother pulling the container image if the digest is the same.