importer: drop some unreachable logic for /opt handling
This simplifies the RPM importing logic by dropping some unused code
for handling the /opt hierarchy.
There is no need to handle this case in the filtering callback, as
paths under /opt have been translated to /usr/lib/opt before
reaching this point.
My understanding is that there was an absolute-vs-relative mixup in
the original implementation, which has been fixed when porting to
Rust. This conditional case was thus left behind but unreachable.