Move all error printing into Rust
Propagate exceptions from the C++ `rpmostree_main()` to the
Rust `main` as a `Result`.
Add a depenency on `termcolor` (also used by `cargo`).
Drop the C code to print colored errors, and move it to Rust
using `termcolor`.  (Which, wow is that not ergonomic, but
hey it's more Rust and less C)