rust/initramfs: Include parent directories in initramfs-etc overlay
Otherwise the kernel will just silently skip over files in the CPIO
which are missing an already extracted parent dir.

It's nice anyway since this also verifies that the path is fully
canonicalized.

There's a minor optimization possible here I didn't feel was worth
doing, which is to pass down `path` into `list_files_recurse` and using
it as a buffer to save a few allocations.

Closes: #2944