Add container image digest and print it in status
I think an important part of this story will be that people
will use the `@sha256:` digested container image references
as a way to track operating system snapshots.  The container
image version isn't always useful because not everyone will
make a container tag for every image.  The ostree commit
hash isn't useful because we can't look it up from the registry.

So let's show the digest front and center.