Add support for modules
Support for modules is something that has come up many times in the
past. Most recently in FCOS:

https://github.com/coreos/fedora-coreos-tracker/issues/767

This commit adds support for both composing with modules on the server
side and package layering modules (or just enabling them) on the client
side.

This doesn't support modules in `rpm-ostree compose extensions` yet,
which is relevant for RHCOS. We can look at adding that in a follow-up.