src: Wrap all throwable CXX calls in try-catch
With the migration to Rust via C++, we've been adding more calls on the
C side which can throw. This in turn means that we lose precious error
context when it finally percolates all the way up.

This patch introduces a new `CXX_TRY` macro (and its variant
`CXX_TRY_VAL`) which is used to wrap every call to C++/Rust code capable
of throwing. This in effect restores error-prefixing for many potential
errors.

Of course, there are exceptions (pun intended). Code which already does
special exception handling was left untouched. C++ code marked as
`noexcept` isn't wrapped (which also means bridged Rust code which don't
return a `Result`).

I'd like to play around a bit more with the macros to see if we can get
rid of having to write out the type in `CXX_TRY_VAL`, but this works for
now.

So essentially after this, everytime we call a throwable C++ function or
CXX-bridged function from a vanilla C code with `GError`-semantics, we
should almost always use `CXX_TRY` or `CXX_TRY_VAL`.

See related discussions in
https://github.com/coreos/rpm-ostree/pull/3026.