countme: Retry hourly on complete failure
This unit is triggered only once par week at a random time. Thus if it
fails completely (no successful counting request sent), it will never
retry for the week thus the system won't be counted in the corresponding
Count Me window. The common temporary and complete failures are usually
due to the network not being available at the time we do the couting.

Add an hourly retry on complete failure only to avoid this specific
case. Couting will still be performed at a random time but shifted by
one or more hours.

In all cases, counting will still only be done once as the logic for
that check is directly performed by the Count Me implementation in
rpm-ostree.

Fixes: https://github.com/coreos/rpm-ostree/issues/3033